Change DAO's count queries from Long to Number to conform to SQL Server

+++ src/main/java/com/github/chipolaris/bootforum/service/ForumService.java
@@ -134,7 +134,7 @@ public class ForumService {
 		
 		// reset forum count
 		SystemInfoService.Statistics systemStat = systemInfoService.getStatistics().getDataObject();
-		systemStat.setForumCount(genericDAO.countEntities(Forum.class));
+		systemStat.setForumCount(genericDAO.countEntities(Forum.class).longValue());
 		
 		return  response;
 	}
